[Decomp][Unit] Cohesion split: extract power/resource accessors to UnitPower.cpp (#218)
Re-applies the mangosthree decomposition onto MangosOne's own TBC 2.4.3
source. Pure code move out of the parent translation unit into the new
cohesively-named file(s); no behaviour change (source diff is deletion-only).
CMake file(GLOB) picks the new TU(s) up automatically. Verified: game.lib
compiles and links clean (MSVC, Release).

Not ported (absent in MangosOne TBC 2.4.3 -- these are
later-expansion methods that do not exist in this branch):
  Unit::GetPowerIndexByClass,Unit::GetPowerTypeByIndex,Unit::GetPower,Unit::GetPowerByIndex,Unit::GetMaxPower,Unit::GetMaxPowerByIndex,Unit::SetPowerByIndex,Unit::SetMaxPowerByIndex,Unit::GetCreateMaxPowers
Unit::GetPowerIndexByClass,Unit::GetPowerTypeByIndex,Unit::GetPower,Unit::GetPowerByIndex,Unit::GetMaxPower,Unit::GetMaxPowerByIndex,Unit::SetPowerByIndex,Unit::SetMaxPowerByIndex,Unit::GetCreateMaxPowers
Ported-from: mangosthree/server@c33ae0e4e736
