+++ src/game/Object/Unit.cpp
@@ -5973,200 +5973,6 @@ void Unit::SetHealthPercent(float percent)
 }
 
 /**
- * @brief Sets a power value while clamping to its maximum.
- *
- * @param power The power type to set.
- * @param val The requested power value.
- */
-void Unit::SetPower(Powers power, uint32 val)
-{
-    if (GetPower(power) == val)
-    {
-        return;
-    }
-
-    uint32 maxPower = GetMaxPower(power);
-    if (maxPower < val)
-    {
-        val = maxPower;
-    }
-
-    SetStatInt32Value(UNIT_FIELD_POWER1 + power, val);
-
-    // group update
-    if (GetTypeId() == TYPEID_PLAYER)
-    {
-        if (((Player*)this)->GetGroup())
-        {
-            ((Player*)this)->SetGroupUpdateFlag(GROUP_UPDATE_FLAG_CUR_POWER);
-        }
-    }
-    else if (((Creature*)this)->IsPet())
-    {
-        Pet* pet = ((Pet*)this);
-        if (pet->isControlled())
-        {
-            Unit* owner = GetOwner();
-            if (owner && (owner->GetTypeId() == TYPEID_PLAYER) && ((Player*)owner)->GetGroup())
-            {
-                ((Player*)owner)->SetGroupUpdateFlag(GROUP_UPDATE_FLAG_PET_CUR_POWER);
-            }
-        }
-
-        // Update the pet's character sheet with happiness damage bonus
-        if (pet->getPetType() == HUNTER_PET && power == POWER_HAPPINESS)
-        {
-            pet->UpdateDamagePhysical(BASE_ATTACK);
-        }
-    }
-}
-
-/**
- * @brief Sets a power maximum and clamps current power if necessary.
- *
- * @param power The power type to update.
- * @param val The new maximum value.
- */
-void Unit::SetMaxPower(Powers power, uint32 val)
-{
-    uint32 cur_power = GetPower(power);
-    SetStatInt32Value(UNIT_FIELD_MAXPOWER1 + power, val);
-
-    // group update
-    if (GetTypeId() == TYPEID_PLAYER)
-    {
-        if (((Player*)this)->GetGroup())
-        {
-            ((Player*)this)->SetGroupUpdateFlag(GROUP_UPDATE_FLAG_MAX_POWER);
-        }
-    }
-    else if (((Creature*)this)->IsPet())
-    {
-        Pet* pet = ((Pet*)this);
-        if (pet->isControlled())
-        {
-            Unit* owner = GetOwner();
-            if (owner && (owner->GetTypeId() == TYPEID_PLAYER) && ((Player*)owner)->GetGroup())
-            {
-                ((Player*)owner)->SetGroupUpdateFlag(GROUP_UPDATE_FLAG_PET_MAX_POWER);
-            }
-        }
-    }
-
-    if (val < cur_power)
-    {
-        SetPower(power, val);
-    }
-}
-
-/**
- * @brief Applies or removes a flat modifier to current power.
- *
- * @param power The power type to modify.
- * @param val The amount to apply or remove.
- * @param apply True to apply the modifier; false to remove it.
- */
-void Unit::ApplyPowerMod(Powers power, uint32 val, bool apply)
-{
-    ApplyModUInt32Value(UNIT_FIELD_POWER1 + power, val, apply);
-
-    // group update
-    if (GetTypeId() == TYPEID_PLAYER)
-    {
-        if (((Player*)this)->GetGroup())
-        {
-            ((Player*)this)->SetGroupUpdateFlag(GROUP_UPDATE_FLAG_CUR_POWER);
-        }
-    }
-    else if (((Creature*)this)->IsPet())
-    {
-        Pet* pet = ((Pet*)this);
-        if (pet->isControlled())
-        {
-            Unit* owner = GetOwner();
-            if (owner && (owner->GetTypeId() == TYPEID_PLAYER) && ((Player*)owner)->GetGroup())
-            {
-                ((Player*)owner)->SetGroupUpdateFlag(GROUP_UPDATE_FLAG_PET_CUR_POWER);
-            }
-        }
-    }
-}
-
-/**
- * @brief Applies or removes a flat modifier to maximum power.
- *
- * @param power The power type to modify.
- * @param val The amount to apply or remove.
- * @param apply True to apply the modifier; false to remove it.
- */
-void Unit::ApplyMaxPowerMod(Powers power, uint32 val, bool apply)
-{
-    ApplyModUInt32Value(UNIT_FIELD_MAXPOWER1 + power, val, apply);
-
-    // group update
-    if (GetTypeId() == TYPEID_PLAYER)
-    {
-        if (((Player*)this)->GetGroup())
-        {
-            ((Player*)this)->SetGroupUpdateFlag(GROUP_UPDATE_FLAG_MAX_POWER);
-        }
-    }
-    else if (((Creature*)this)->IsPet())
-    {
-        Pet* pet = ((Pet*)this);
-        if (pet->isControlled())
-        {
-            Unit* owner = GetOwner();
-            if (owner && (owner->GetTypeId() == TYPEID_PLAYER) && ((Player*)owner)->GetGroup())
-            {
-                ((Player*)owner)->SetGroupUpdateFlag(GROUP_UPDATE_FLAG_PET_MAX_POWER);
-            }
-        }
-    }
-}
-
-/**
- * @brief Registers or unregisters an aura in the proc-trigger-damage list.
- *
- * @param aura The aura to add or remove.
- * @param apply True to add the aura; false to remove it.
- */
-void Unit::ApplyAuraProcTriggerDamage(Aura* aura, bool apply)
-{
-    AuraList& tAuraProcTriggerDamage = m_modAuras[SPELL_AURA_PROC_TRIGGER_DAMAGE];
-    if (apply)
-    {
-        tAuraProcTriggerDamage.push_back(aura);
-    }
-    else
-    {
-        tAuraProcTriggerDamage.remove(aura);
-    }
-}
-
-/**
- * @brief Gets the default base value for a power type.
- *
- * @param power The power type to query.
- * @return The created base power value.
- */
-uint32 Unit::GetCreatePowers(Powers power) const
-{
-    switch (power)
-    {
-        case POWER_HEALTH:      return 0;                   // is it really should be here?
-        case POWER_MANA:        return GetCreateMana();
-        case POWER_RAGE:        return POWER_RAGE_DEFAULT;
-        case POWER_FOCUS:       return (GetTypeId() == TYPEID_PLAYER || !((Creature const*)this)->IsPet() || ((Pet const*)this)->getPetType() != HUNTER_PET ? 0 : POWER_FOCUS_DEFAULT);
-        case POWER_ENERGY:      return POWER_ENERGY_DEFAULT;
-        case POWER_HAPPINESS:   return (GetTypeId() == TYPEID_PLAYER || !((Creature const*)this)->IsPet() || ((Pet const*)this)->getPetType() != HUNTER_PET ? 0 : POWER_HAPPINESS_DEFAULT);
-        default: break; //MAX_POWERS and POWERS_ALL probably should not belong to the enum Powers to do not require the default: case
-    }
-
-    return 0;
-}
-
-/**
  * @brief Adds the unit to the world and initializes world-dependent helpers.
  */
 void Unit::AddToWorld()
